xso: Make ChildTextMap accept either a tag or an XSO
ChildTextMap now can be constructed from either an XSO or a tag.
If a tag is given a matching XSO is constructed on the fly.
